Update on User Model, Scholarship, Events and user relationship (#16)
* Changing primary keys to id

* Family detail feature update

* Event and scholarship edit function

* Display events feature added

+++ Backend/src/main/java/com/example/falsarii/backend/repository/DonateToScholarshipsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.falsarii.backend.model.DonateToScholarships;
 
 @Repository
-public interface DonateToScholarshipsRepository extends JpaRepository<DonateToScholarships, Integer> {
+public interface DonateToScholarshipsRepository extends JpaRepository<DonateToScholarships, Long> {
 	
 }
